+++ src/js/models/lifetime-model.js
@@ -1,0 +1,22 @@
+'use strict';
+
+/*
+  lifetimeModel is a class that stores lifetime benefit values for each claiming age.
+  */
+var lifetimeModel = {
+  values: {
+    'age62': 1,
+    'age63': 1,
+    'age64': 1,
+    'age65': 1,
+    'age66': 1,
+    'age67': 1,
+    'age68': 1,
+    'age69': 1,
+    'age70': 1
+  }
+};
+
+
+module.exports = lifetimeModel;
+