+++ src/js/wizards/update-model.js
@@ -1,13 +1,16 @@
 'use strict';
 
 var benefitsModel = require( '../models/benefits-model' );
+var lifetimeModel = require( '../models/lifetime-model' );
 
 var update = {
 
   benefits: function( prop, val ) {
     benefitsModel.values[prop] = val;
   },
-
+  lifetime: function( prop, val ) {
+    lifetimeModel.values[prop] = val;
+  },
   processApiData: function( resp ) {
     var data = resp.data,
         fullAge = Number( data['full retirement age'].substr( 0, 2 ) );
@@ -19,6 +22,9 @@ var update = {
         var prop = i.replace( ' ', '' );
         update.benefits( prop, val );
       }
+    } );
+    $.each( resp.data.lifetime, function( prop, val ) {
+        update.lifetime( prop, val );
     } );
     update.benefits( 'currentAge', resp.current_age );
     update.benefits( 'past_fra', resp.past_fra );