+++ CHANGELOG.md
@@ -4,6 +4,8 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 ## Unreleased
 - Deleted unneeded grunt packages
 - Added `.npmrc` and `npm-shrinkwrap.json` files
+- Added lifetime calculations to api and removed from js
+- Removed redundant cf-icon CSS code
 
 ## 0.4.62
 - Changed the url namespace for the 'about' page to be specific to retirement